+++ Changelog.md
@@ -156,6 +156,7 @@ and this project follows to [Ragnarök Versioning Convention](https://gist.githu
 - Fixed melee attack and heavy melee attack being switched
 - Fixed inputs not updating if you switched to another hotbar slot while holding the input
 - Fixed oreDictionary not working properly under the craftmappings
+- Fixed the height of the Ammo Counter being 1 pixel too high
 
 ### Removed